requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.1
langgraph>=0.0.20
google-generativeai>=0.8.0
comet-llm>=0.0.1
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
elevenlabs>=0.2.0
echo-ai-sdk>=1.0.0
